+++ tools/dartium/build.py
@@ -12,7 +12,6 @@ ALL_TARGETS = [
     'content_shell',
     'chrome',
     'blink_tests',
-    'pkg_packages',
 ]
 
 def main():